+++ packages/core/src/decorators/target.ts
@@ -8,6 +8,7 @@ export type TargetType = {
 export function target() {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   return (ctor: any, key: string) => {
+    Object.defineProperty(Object.getPrototypeOf(ctor), key, { configurable: true, value: null });
     const store = new Store(ctor, 'target');
     store.add({ key, multiple: false });
   };
@@ -16,6 +17,7 @@ export function target() {
 export function targets() {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   return (ctor: any, key: string) => {
+    Object.defineProperty(Object.getPrototypeOf(ctor), key, { configurable: true, value: [] });
     const store = new Store(ctor, 'target');
     store.add({ key, multiple: true });
   };