+++ packages/core/src/target.ts
@@ -2,11 +2,11 @@ import SetMap from './data_structures/set_map';
 import type { TargetType } from './decorators/target';
 import type ImpulseElement from './element';
 import { capitalize } from './helpers/string';
-import TokenListObserver, { Token } from './observers/token_list_observer';
+import { Token, TokenListObserver, TokenListObserverDelegate } from './observers/token_list_observer';
 import Scope from './scope';
 import Store from './store';
 
-export default class Target {
+export default class Target implements TokenListObserverDelegate {
   private store: Store;
   private scope: Scope;
   private targetsByKey: SetMap<string, Element>;
@@ -23,7 +23,6 @@ export default class Target {
     if (!this.tokenListObserver) {
       this.tokenListObserver = new TokenListObserver(this.instance, 'data-target', this);
       this.tokenListObserver.start();
-      this.initializeTargets();
     }
   }
 
@@ -54,7 +53,7 @@ Learn more about the @targets() decorator: https://ambiki.github.io/impulse/refe
       );
     }
 
-    this.defineProperty(key, this.isKeyMultiple(key) ? this.targetsByKey.getValuesForKey(key) : element);
+    this.defineProperty(key, this.isKeyMultiple(key) ? targets : element);
     this.invokeCallback(key, element, 'connected');
   }
 
@@ -66,18 +65,6 @@ Learn more about the @targets() decorator: https://ambiki.github.io/impulse/refe
     this.invokeCallback(key, element, 'disconnected');
     // Update property after invoking callback.
     this.defineProperty(key, this.isKeyMultiple(key) ? this.targetsByKey.getValuesForKey(key) : null);
-  }
-
-  private initializeTargets() {
-    for (const key of this.keys) {
-      const targets = this.scope.findTargets(`[data-target~="${this.identifier}.${key}"]`);
-      // If no targets were found, define the property as null or an empty array.
-      if (!targets.length) {
-        this.defineProperty(key, this.isKeyMultiple(key) ? [] : null);
-      }
-
-      targets.forEach((target) => this.tokenListObserver?.elementConnected(target));
-    }
   }
 
   private destroyTargets() {